refactor(profile): move ProfileFormContext to hook layer, restore Fields.tsx generic reusability
Address AI review architecture/correctness findings:
- Move ProfileFormContext out of profileSchema.ts (schema layer should stay
  free of react-hook-form's UI types) and export it from useEditProfileForm.ts
  instead, where all consumers now import it from.
- Give TextField/TextareaField a separate TTransformedValues generic so the
  shared field components stay reusable for forms whose Zod schema uses
  .default()/transforms (input and output types differing), bridging to the
  shadcn FormField wrapper's fixed Control<T, any, T> signature with a narrow
  internal cast.
