+++ src/app/profile/[pageUserId]/edit/container.test.tsx
@@ -131,8 +131,8 @@ vi.mock('@/components/profile/edit/LinkSection', () => ({
 }));
 
 import * as useEditProfileFormModule from '@/hooks/user/profile/useEditProfileForm';
+import { ProfileFormContext } from '@/hooks/user/profile/useEditProfileForm';
 import { MENTOR_ONBOARDING_KEY } from '@/lib/routes';
-import { ProfileFormContext } from '@/schemas/profileSchema';
 import type { TagCatalogsByBucket } from '@/services/profile/tagCatalog';
 import { MentorProfileVO } from '@/services/profile/user';
