+++ src/schemas/profileSchema.ts
@@ -1,4 +1,3 @@
-import { UseFormReturn } from 'react-hook-form';
 import * as z from 'zod';
 
 //--------------------------------------------------
@@ -164,12 +163,6 @@ export type ProfileFormValues = z.infer<
 
 export type ProfileFormInput = z.input<
   ReturnType<typeof createProfileFormSchema>
->;
-
-export type ProfileFormContext = UseFormReturn<
-  ProfileFormInput,
-  unknown,
-  ProfileFormValues
 >;
 
 export const defaultValues: ProfileFormValues = {
